Improved: D2X-XL will now generally seed the random generator with 32 bit random values (using SDL_Ticks() - elapsed time since program start in ms)
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11734 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/descent.h
@@ -1230,7 +1230,7 @@ class CApplicationStates {
 		char *szCurrentMissionFile;
 		tObjTransformation playerPos;
 		short nPlayerSegment;
-		short nRandSeed;
+		uint nRandSeed;
 		CCheatStates cheats;
 	};
